Move GeolocationDbUpdater from CLI to Core module
